requirements.txt
requests
python-dateutil
pytest
pytz
python-dotenv
pyinstaller
